Revert "Arraged file in proper folder"
This reverts commit 96aa4b9c18ac349d0f1243f6702ec0892cada299.
