--- ClassesAndObjects/008PointersToMember.cpp
@@ -1,11 +1,0 @@
-#include <iostream>
-using namespace std;
-
-class A{
-    private:
-    int m;
-    public:
-    void show(void);
-};
-
-// int A::* ip = &A :: m;
